+++ wsgidav/default_conf.py
@@ -14,10 +14,11 @@ r"""
 
 Default confguration.
 """
-from wsgidav.debug_filter import WsgiDavDebugFilter
+# from wsgidav.mw.debug_filter import WsgiDavDebugFilter
 from wsgidav.dir_browser import WsgiDavDirBrowser
 from wsgidav.error_printer import ErrorPrinter
 from wsgidav.http_authenticator import HTTPAuthenticator
+from wsgidav.mw.cors import Cors
 from wsgidav.request_resolver import RequestResolver
 
 __docformat__ = "reStructuredText"
@@ -38,7 +39,8 @@ DEFAULT_CONFIG = {
     "hotfixes": {
         "emulate_win32_lastmod": False,  # True: support Win32LastModifiedTime
         "re_encode_path_info": True,  # (See issue #73)
-        "unquote_path_info": False,  # See issue #8, #228
+        "unquote_path_info": False,  # (See issue #8, #228)
+        # "treat_root_options_as_asterisk": False, # Hotfix for WinXP / Vista: accept 'OPTIONS /' for a 'OPTIONS *'
         # "win_accept_anonymous_options": False,
         # "winxp_accept_root_share_login": False,
     },
@@ -46,7 +48,8 @@ DEFAULT_CONFIG = {
     "mutable_live_props": [],
     "lock_storage": True,  # True: use LockManager(lock_storage.LockStorageDict)
     "middleware_stack": [
-        WsgiDavDebugFilter,
+        # WsgiDavDebugFilter,
+        Cors,
         ErrorPrinter,
         HTTPAuthenticator,
         WsgiDavDirBrowser,  # configured under dir_browser option (see below)