+++ wsgidav/fs_dav_provider.py
@@ -41,7 +41,7 @@ class FileResource(DAVNonCollection):
     """
 
     def __init__(self, path, environ, file_path):
-        super(FileResource, self).__init__(path, environ)
+        super().__init__(path, environ)
         self._file_path = file_path
         self.file_stat = os.stat(self._file_path)
         # Setting the name from the file path should fix the case on Windows
@@ -175,7 +175,7 @@ class FolderResource(DAVCollection):
     """
 
     def __init__(self, path, environ, file_path):
-        super(FolderResource, self).__init__(path, environ)
+        super().__init__(path, environ)
         self._file_path = file_path
         self.file_stat = os.stat(self._file_path)
         # Setting the name from the file path should fix the case on Windows
@@ -355,17 +355,21 @@ class FolderResource(DAVCollection):
 # FilesystemProvider
 # ========================================================================
 class FilesystemProvider(DAVProvider):
-    def __init__(self, root_folder_path, *, readonly=False):
-        # Expand leading '~' as user home dir; expand %VAR%, $Var, ..
-        root_folder_path = os.path.expandvars(os.path.expanduser(root_folder_path))
-        root_folder_path = os.path.abspath(root_folder_path)
-        if not root_folder_path or not os.path.exists(root_folder_path):
-            raise ValueError("Invalid root path: {}".format(root_folder_path))
+    def __init__(self, root_folder, *, readonly=False, shadow=None):
+        # and resolve relative to config file
+        # root_folder = os.path.expandvars(os.xpath.expanduser(root_folder))
+        root_folder = os.path.abspath(root_folder)
+        if not root_folder or not os.path.exists(root_folder):
+            raise ValueError("Invalid root path: {}".format(root_folder))
 
-        super(FilesystemProvider, self).__init__()
+        super().__init__()
 
-        self.root_folder_path = root_folder_path
+        self.root_folder_path = root_folder
         self.readonly = readonly
+        if shadow:
+            self.shadow = {k.lower(): v for k, v in shadow.items()}
+        else:
+            self.shadow = {}
 
     def __repr__(self):
         rw = "Read-Write"
@@ -374,6 +378,28 @@ class FilesystemProvider(DAVProvider):
         return "{} for path '{}' ({})".format(
             self.__class__.__name__, self.root_folder_path, rw
         )
+
+    def _resolve_shadow_path(self, path, environ, file_path):
+        """File not found: See if there is a shadow configured."""
+        shadow = self.shadow.get(path.lower())
+        # _logger.info(f"Shadow {path} -> {shadow} {self.shadow}")
+        if not shadow:
+            return False, file_path
+
+        err = None
+        method = environ["REQUEST_METHOD"].upper()
+        if method not in ("GET", "HEAD", "OPTIONS"):
+            err = f"Shadow {path} -> {shadow}: ignored for method {method!r}."
+        elif os.path.exists(file_path):
+            err = f"Shadow {path} -> {shadow}: ignored for existing resource {file_path!r}."
+        elif not os.path.exists(shadow):
+            err = f"Shadow {path} -> {shadow}: does not exist."
+
+        if err:
+            _logger.warning(err)
+            return False, file_path
+        _logger.info(f"Shadow {path} -> {shadow}")
+        return True, shadow
 
     def _loc_to_file_path(self, path, environ=None):
         """Convert resource path to a unicode absolute file path.
@@ -387,7 +413,8 @@ class FilesystemProvider(DAVProvider):
 
         path_parts = path.strip("/").split("/")
         file_path = os.path.abspath(os.path.join(root_path, *path_parts))
-        if not file_path.startswith(root_path):
+        is_shadow, file_path = self._resolve_shadow_path(path, environ, file_path)
+        if not file_path.startswith(root_path) and not is_shadow:
             raise RuntimeError(
                 "Security exception: tried to access file outside root: {}".format(
                     file_path