+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Drop Python 2 support
 - Drop support for Python syntax in config files (wsgidav.conf)
 - Drop support for Microsoft Web Folders (option `dir_browser.ms_mount`).
+- CORS support
+- Provider root paths are relative to configuration file
 - DAVCollection, DAVNonCollection, DAVProvider are now ABCs.
 - API enforces some named keyword args (`..., *, ...`)
 - Rename option `lock_manager` to `lock_storage`
@@ -330,4 +332,4 @@ This release contains **BREAKING CHANGES!**
 
 ## Until 0.4.0 alpha
 
-See https://github.com/mar10/wsgidav/blob/master/docs/changelog04.md
+See https://github.com/mar10/wsgidav/blob/master/docs/source/changelog04.md